[IMP] selection: don't dispatch ACTIVATE_SHEET at START
The selection plugin currently dispatch an ACTIVATE_SHEET command at START.
This means that other UI plugins cannot use the START command to initialize
themselves, because by the time they receive it, ACTIVATE_SHEET is
already dispatched, and the viewports are recomputed. But recomputing the
viewports need information such as the size/visibility of the headers.

Might be worth it in the future to add a method similar to the `import`
of core plugins, but for UI plugins, so they can be initialized before
handling commands.

Part of task 3240138