[IMP] Menu: wait to close submenu
Task Description

When overing an item in a menu, we open the corresponding submenu
(if present). Sometimes, when going "too fast" to go from the
item to its submenu, we mistakingly over another item in the same
parent menu, closing then the submenu. This task aims to add a
delay to close a submenu and open another submenu, in the same way
as Google Sheet.

Related Task

Task: 2930428