[TMP] components: remove old static props declaration
To get in line with the Js framework guidelines, and to have something
more modern, we will replace all the old static props declaration
(`Grid.props = `) by more modern syntax (`static props = `).

This commit is a temporary commit. We cannot do it ATM because of a bug
in Owl (https://github.com/odoo/owl/issues/1448), which prevent us
from using `Function` as the type of a props. Changeed all the
`Function` to `any` waiting for the fix to come to owl. We probably
don't want to merge it ATM.