[IMP] topbar: allow to insert checkbox in a cell
Task Description

Now that we have the datavalidation, this PR aims to add a simple
item in the Insert menu to let the user directly insert a checkbox
in a cell without having to use the Data-validation side-pannel.
To mimic Google sheet behavior, we are now also able to remove a
checkbox from a cell in the same way we can delete the content of a
cell (e.g. with 'delete' or 'backspace' key), and copying an empty
cell to a cell with a dataValidation rule remove this rule.

Related Task

Task: 3175856